测试队列

+++ application/admin/controller/job/Demo.php
@@ -3,6 +3,7 @@ namespace app\admin\controller\job;
 
 use think\Db;
 use think\queue\Job;
+use redis\Redis;
 class Demo
 {
     /**
@@ -11,11 +12,17 @@ class Demo
      * @param array|mixed    $data     发布任务时自定义的数据
      */
     public function fire(Job $job,$data){
+        $redis = new Redis();
         $isJobDone = $this->doHelloJob($data);
 
         if ($isJobDone) {
+            $isok = $redis->get('gotostatus');
+
+            if($isok==3){
+               $job->delete();
+            }
             //如果任务执行成功， 记得删除任务
-            $job->delete();
+            
             print("<info>Hello Job has been done and deleted"."</info>\n");
         }else{
             if ($job->attempts() > 3) {
@@ -30,14 +37,39 @@ class Demo
      * 根据消息中的数据进行实际的业务处理
      * @param array|mixed    $data     发布任务时自定义的数据
      * @return boolean                 任务执行的结果
+     * 0-等待继续 1-确认继续 2-
      */
     private function doHelloJob($data) {
         // 根据消息中的数据进行实际的业务处理...
+        $redis = new Redis();
+        $goto = $redis->get('gotostatus');
+        if($goto!=1){return true;}
+        $neednum = $redis->get('has_create');
+        if(!$neednum){return true;}
+        if($neednum-200<0){
+            $now_need = 200-(abs($neednum-200));
+            $set_need = $now_need;
+            $neednum = $redis->del('has_create');
+            $is_ok = 2;
+        }else{
+            $now_need = $neednum-200;
+            $set_need = 200;
+            $neednum = $redis->set('has_create',$now_need);
+            $is_ok = 0;
+        }
+                
+        $jobData = [
 
-     
-        $name = $data['name'];
-        $res = Db::table('user')->insertAll($name);
-
+        ];
+        for ($i=0; $i <$set_need ; $i++) { 
+            # code...
+            $jobData[$i] = ['name'=>'GNLEON'.time()];            
+        }
+        // $name = $data['name'];
+        $res = Db::table('user')->insertAll($jobData);
+        if($res){
+            $redis->set('gotostatus',$is_ok);
+        }
         return $res;
     }
 }
